Use new `spfun/studentT.ccPv` for exact small sample pValues(autocorrs). Exact can easily make 2X deltas in small sample false-positive rates which can easily take 0.04 -> 0.08, for example.
As part of this, change to report the actual p-Value not its complement.

Also update the help strings to fit in 80-columns with cligen defaults
and fix a long-standing annoyance of *[cs] "leaving open" SGR region.

This nullifies the speed-up of 76892b8ffb93b4b858f1da927313b1fb3d3a683d,
of course { but leave that code to support the doubtless zillion users
relying on the `import`. ;-) }